+++ src/wasm-s-parser.h
@@ -285,7 +285,7 @@ private:
   Expression* makeTupleMake(Element& s);
   Expression* makeTupleExtract(Element& s);
   Expression* makeCallRef(Element& s, bool isReturn);
-  Expression* makeI31New(Element& s);
+  Expression* makeRefI31(Element& s);
   Expression* makeI31Get(Element& s, bool signed_);
   Expression* makeRefTest(Element& s);
   Expression* makeRefCast(Element& s);