+++ src/wasm2js.h
@@ -2275,7 +2275,7 @@ Ref Wasm2JSBuilder::processFunctionBody(Module* m,
       unimplemented(curr);
       WASM_UNREACHABLE("unimp");
     }
-    Ref visitI31New(I31New* curr) {
+    Ref visitRefI31(RefI31* curr) {
       unimplemented(curr);
       WASM_UNREACHABLE("unimp");
     }